requirements.txt
pystray==0.19.5
Pillow==10.2.0
psutil==5.9.8
python-dotenv==1.0.1
setuptools
